backend/requirements.txt
fastapi
uvicorn
python-multipart
python-dotenv
langchain
langchain-community
scikit-learn
numpy
google-generativeai
azure-cognitiveservices-vision-computervision
pypdf
python-docx
reportlab
pillow
pydantic
itsdangerous
sqlalchemy
bcrypt
PyJWT
